+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 mail-parser 0.6.2
 ================================
 - Base64/QuotedPrintable decoding optimizations.
+- Added "ludicrous mode" Cargo option to use some unsafe code for additional performance.
   
 mail-parser 0.6.1
 ================================